+++ digital_receipt/lib/utils/theme_manager.dart
@@ -118,8 +118,8 @@ class ThemeManager {
       data: ThemeData(
           cardColor: DarkMode.appBarColor,
           brightness: Brightness.dark,
-          scaffoldBackgroundColor: Colors.amberAccent,
-          //scaffoldBackgroundColor: DarkMode.backgroundColor,
+          //scaffoldBackgroundColor: Colors.amberAccent,
+          scaffoldBackgroundColor: DarkMode.backgroundColor,
           primaryColor: DarkMode.primaryColor,
           primaryColorDark: DarkMode.appBarColor,
           accentColor: Color(0xFF25CCB3),
